LayoutTests/accessibility/gtk/text-at-offset-embedded-objects-expected.txt

This is one line containing two  embedded objects  in the middle.

This tests the ability to get element text for different granularities and offsets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS content.characterAtOffset(32) is '￼, 32, 33'
PASS content.wordAtOffset(32) is 'two ￼ , 28, 34'
PASS content.characterAtOffset(33) is ' , 33, 34'
PASS content.wordAtOffset(33) is 'two ￼ , 28, 34'
PASS content.characterAtOffset(34) is 'e, 34, 35'
PASS content.wordAtOffset(34) is 'embedded , 34, 43'
PASS content.characterAtOffset(42) is ' , 42, 43'
PASS content.wordAtOffset(42) is 'embedded , 34, 43'
PASS content.characterAtOffset(51) is '￼, 51, 52'
PASS content.wordAtOffset(51) is 'objects ￼ , 43, 53'
PASS content.characterAtOffset(52) is ' , 52, 53'
PASS content.wordAtOffset(52) is 'objects ￼ , 43, 53'
PASS content.characterAtOffset(53) is 'i, 53, 54'
PASS content.wordAtOffset(53) is 'in , 53, 56'
PASS content.characterAtOffset(55) is ' , 55, 56'
PASS content.wordAtOffset(55) is 'in , 53, 56'
PASS successfullyParsed is true

TEST COMPLETE
